requirements.txt
crewai
python-dotenv
pydantic
pandas
pyyaml 